requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openai==1.35.3
orjson==3.10.5
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
setuptools==65.5.0
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.12.1
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
numpy

# aiohttp==3.8.5
# aiosignal==1.3.1
# annotated-types==0.5.0
# anyio==3.7.1
# async-timeout==4.0.2
# attrs==23.1.0
# autopep8==2.0.2
# cachetools==5.3.3
# certifi==2023.7.22
# cffi==1.15.1
# charset-normalizer==3.2.0
# click==8.1.6
# distro==1.9.0
# dnspython==2.4.1
# email-validator==2.0.0.post2
# exceptiongroup==1.1.2
# fastapi==0.100.1
# frozenlist==1.4.0
# google-api-core==2.18.0
# google-auth==2.29.0
# google-cloud-core==2.4.1
# google-cloud-storage==2.16.0
# google-crc32c==1.5.0
# google-resumable-media==2.7.0
# googleapis-common-protos==1.63.0
# h11==0.14.0
# httpcore==0.17.3
# httptools==0.6.0
# httpx==0.24.1
# idna==3.4
# itsdangerous==2.1.2
# Jinja2==3.1.2
# MarkupSafe==2.1.3
# multidict==6.0.4
# numpy==1.25.1
# openai==1.17.1
# orjson==3.9.2
# proto-plus==1.23.0
# protobuf==4.25.3
# pyasn1==0.6.0
# pyasn1_modules==0.4.0
# pycodestyle==2.10.0
# pycparser==2.21
# pydantic==2.1.1
# pydantic-extra-types==2.0.0
# pydantic-settings==2.0.2
# pydantic_core==2.4.0
# python-dotenv==1.0.0
# python-multipart==0.0.6
# PyYAML==6.0.1
# requests==2.31.0
# rsa==4.9
# sniffio==1.3.0
# soundfile==0.12.1
# starlette==0.27.0
# tomli==2.0.1
# tqdm==4.65.0
# typing_extensions==4.7.1
# ujson==5.8.0
# urllib3==2.0.4
# uvicorn==0.23.1
# uvloop==0.17.0
# watchfiles==0.19.0
# websockets==11.0.3
# yarl==1.9.2
